Fix usage of DESTDIR.
When DISTDIR is not specified "$(DESTDIR)/@lispdir@/lookup" is
expanded something like "//usr/local/share/emacs/sitelisp/lookup".  On
Unix it doesn't cause problem because "//" is same as "/". But on
Windows it is interpreted as specifying remote file path and may
result in error. So fix problem by changing it to
"$(DESTDIR)@lispdir@/lookup"